[INCOMPLETE] Allow commas within hts_parse_decimal()
Allow commas to be used to separate thousands in fields parsed by
hts_parse_decimal().  The inend parameter is to support parsing prefixes
of text, such as 1000-2000 (where inend would be instr+4) and such as beg
in hts_parse_reg(), but perhaps it's better just to assume that the range
hyphen in 1,230e-1-125 would always be detected as the number terminator.

Needs further work in hts_parse_reg(), which no longer needs to strip
commas itself, and whose detection of valid (scientific) number characters
may need further work.